% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.db.R
\name{code.db}
\alias{code.db}
\title{code.db}
\usage{
head(code.db('repos/robertzk/3chessengine',verbose=T))
}
\arguments{
\item{repo.path}{- The api.github.com path to a single repo.}

\item{.tail}{- I honestly don't remember my reasoning here. It's just the
last end of the url. I can't see why the user would want to change it, but
YOLO.}

\item{verbose}{- Binary flag useful for troubleshooting. Will print out each
file path checked to terminal if true. Lets you know which files are generating
errors.}
}
\value{
returns a single data frame listing all the contents of the repo.
}
\description{
Recursive function that looks through a repo for all potential
scripts, and returns them as a table, including several relevant properties,
but more importantly the raw text. The objective is to have all repo contents
as a data frame, so the user can analyze it like any other data set.
}

