% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffs.db.R
\name{diffs.db}
\alias{diffs.db}
\title{diffs.db}
\usage{
df <- diffs.db(
  'repos/robertzk/3chessengine'
  ,'6591146d25ccb08eae904607f59ac2207e7b7397'
  ,'master'
  ,'unified'
 )
df$diff.url
head(df$diff.data , 5)
}
\arguments{
\item{repo.path}{- The path to the repo}

\item{sha.start}{- The full sha key for the commit you wish to start on.}

\item{sha.end}{- The full sha key for the commit you wish to end on. Defaults
to master.}

\item{type}{- 'unified' or 'split'. Entering any other option here will break
the url, so don't do that.}
}
\value{
diff.url - The path to the full change file. Just put it in your
browser.

diff.data - A table with the file name, properties, and the column
'patch', which is the full text of the change. Use this when you want to
select subsets of changes for review - e.g. excluding readmes and .png files.
}
\description{
Creates a single table and a url. The url will direct you to
the diff file, but you have to copy it into the browser yourself. (I'd like to
get the full file instead, but I'm already over and need to get this out). The
table returned is each changed file, some properties of the file, and the
text including the changes.
}

