% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.git.R
\name{get.git}
\alias{get.git}
\title{get.git}
\usage{
get.git(path, two.fac = 0, user = getOption("GITHUB_USER"),
  pwd = getOption("GITHUB_TOKEN"))
}
\arguments{
\item{path}{- The raw.githubusercontent.com path to the file. Should be similar
to the path used by api.github.com}
}
\value{
Function returns the code as a character string
}
\description{
Function returns the raw-text code in a specific GitHub file.
}
\examples{
get.git("robertzk/3chessengine/master/.gitignore")
get.git("robertzk/3chessengine/master/README.md")
get.git("robertzk/3chessengine/master/bower.json")
}

