% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_commits.R
\name{get_all_commits}
\alias{get_all_commits}
\title{Get all commits}
\usage{
get_all_commits(git_repo)
}
\arguments{
\item{git_repo}{path to git repo}
}
\value{
data frame
}
\description{
Get all commits
}
\examples{
NULL
}
