% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_commits_as_chunks.R
\name{get_all_commits_as_chunks}
\alias{get_all_commits_as_chunks}
\title{Get all commits as chunks}
\usage{
get_all_commits_as_chunks(git_repo)
}
\arguments{
\item{git_repo}{path to git repo}
}
\value{
data frame
}
\description{
Get all commits as chunks
}
\examples{
NULL
}
