% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_commit_as_chunks.R
\name{get_single_commit_as_chunks}
\alias{get_single_commit_as_chunks}
\title{Get single commit as chunks from repo}
\usage{
get_single_commit_as_chunks(git_repo, commit_id = 10)
}
\arguments{
\item{git_repo}{path to repo}

\item{commit_id}{numeric commit id}
}
\value{
data frame
}
\description{
Get single commit as chunks from repo
}
\examples{
NULL
}
