\name{GitR-package}
\alias{GitR-package}
\alias{GitR}
\docType{package}
\title{
\packageTitle{GitR}
}
\description{
\packageDescription{GitR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GitR}
\packageIndices{GitR}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{GitR}

Maintainer: \packageMaintainer{GitR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
