% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.path.R
\name{install.path}
\alias{install.path}
\title{install.path}
\usage{
install.path('users/robertzk')
}
\arguments{
\item{path}{- The api.github.com path to a user account or organization, or
anywhere where repos are stored.}
}
\value{
Null. But it sure will print out a ton of diagnostic stuff, so get
ready. There's usually a bunch of dependency bullshit in a user's list of repos,
so if you get errors on packages that are legit R repositories, just keep
running this function until the error list stabilizes
}
\description{
Installs every repo at a particular user's path, skipping over
any repo that is not an R repo.
}

