% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.db.R
\name{issues.db}
\alias{issues.db}
\title{issues.db}
\usage{
issues.db(repo.path)
}
\arguments{
\item{repo.path}{- The api.github.com path to a single repo.}
}
\description{
Creates an issues and a comments table from the path to a
specific Github repo. Tables are joinable on issues$id = comments$issue_id,
and the relationship is 1-many.
}
\examples{
issues.db("repos/avantcredit/analytics-partners")
}

