% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.repos.R
\name{list.repos}
\alias{list.repos}
\title{list.repos}
\usage{
list.repos('repos/robertzk/3chessengine')
}
\arguments{
\item{path}{- A user or org path, e.g. 'users/christiantillich'}
}
\value{
Returns a list of the full paths to any visible repo on the account.
}
\description{
Lists all the repos at a given api.github.com path.
}

