% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.github.path.R
\name{show.github.path}
\alias{show.github.path}
\title{show.github.path}
\usage{
show.github.path(path, two.fac = 0, user = getOption("GITHUB_USER"),
  pwd = getOption("GITHUB_TOKEN"))
}
\arguments{
\item{path}{- Not the full path. Just the parts after api.github.com/, e.g.
'user/christiantillich/repos', as a character string.}

\item{two.fac}{- Added for two-factor authentication. I'm using a PAT, though
so I've never actually tested whether this works.}

\item{user}{- defaults to the global variable GITHUB_USER. Please add this
to your global variables in your .RProfile}

\item{pwd}{- defaults to the global variable GITHUB_TOKEN. Please create a
valid token via https://github.com/settings/tokens and add this to your
.RProfile. Alternatively, you can just use your password. However, because
we're locked into 2-factor authentication here, I haven't tested this.}
}
\description{
Function returns the content of a github path
}

