% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.db.R
\name{issues.db}
\alias{issues.db}
\title{issues.db}
\usage{
issues.db('repos/robertzk/3chessengine')
}
\arguments{
\item{repo.path}{- The api.github.com path to a single repo.}
}
\value{
Function returns a list of tables, "issues" and "comments"
}
\description{
Creates an issues and a comments table from the path to a
specific Github repo. Tables are joinable on issues$id = comments$issue_id,
and the relationship is 1-many.
}

